add: Grumpkin curve to the workspace
Adds support for the Grumpkin curve in arkworks. Grumpkin is a SW curve
which forms a cycle with Bn254.
This allows to use the Bn254-Grumpkin cycle in contexts like folding
(where for instance is needed by https://github.com/privacy-scaling-explorations/folding-schemes/pull/11
so that we can work on on-chain verification for folding).

This also allows to use this cycle for Halo2 with Hyperplonk for example
(as the 2Adacity of Grumpkin is 1 so KZG doesn't work).